+++ Application/FileConverter/Diagnostics/Debug.cs
@@ -21,7 +21,7 @@ namespace FileConverter.Diagnostics
         {
             Debug.mainThreadId = Thread.CurrentThread.ManagedThreadId;
 
-            string path = PathHelpers.GetUserDataFolderPath();
+            string path = FileConverterExtension.PathHelpers.GetUserDataFolderPath;
 
             // Delete old diagnostics folder (1 day).
             DateTime expirationDate = DateTime.Now.Subtract(new TimeSpan(1, 0, 0, 0));
@@ -88,6 +88,14 @@ namespace FileConverter.Diagnostics
             }
 
             diagnosticsData.Log(log);
+        }
+
+        public static void Assert(bool condition)
+        {
+            if (!condition)
+            {
+                LogError("Assertion failed");
+            }
         }
 
         public static void Assert(bool condition, string message)